feat: change usePasswordInput and PasswordInput API
